pyproject.toml
[project]
name = "task-tracker"
version = "0.1.0"
description = "Приложение для распознавания переговоров и их суммаризации, а также отслеживания статусов задач в рабочих чатах."
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "aio-pika>=9.5.8",
    "pydantic>=2.12.3",
    "pydantic-settings>=2.11.0",
]

[project.optional-dependencies]

api-gateway = [
    "fastapi>=0.120.1",
    "uvicorn>=0.38.0",
]

canban-connectors = [
    "py-trello>=0.20.1",
]

similarity = [

]

speech-transcriptor = [
    "nemo>=7.2.5",
    "nemo-toolkit[asr]>=2.1.0",
    "noisereduce>=3.0.3",
    "numpy>=2.2.6",
    "openai-whisper>=20250625",
    "pandas>=2.3.3",
    "pyannote-audio>=4.0",
    "pydub>=0.25.1",
    "scipy>=1.16.3",
    "torchaudio>=2.8.0,<2.9.0",
    "whisperx>=3.4.3",
]

text-summarizer = [
    "openai>=2.8.1",
    "vllm>=0.11.2",
]

user-connectors = [

]
